pyproject.toml
[tool.poetry]
name = "receipt-processor-challenge"
version = "0.1.0"
description = ""
authors = ["Michael Fernandez <michaelf2465@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.9.6"
fastapi = {extras = ["standard"], version = "^0.115.8"}
sqlalchemy = "^2.0.38"
coloredlogs = "^15.0.1"
pyyaml = "^6.0.2"
pydantic-settings = "^2.7.1"
sqlalchemy-utils = "^0.41.2"
sqlmodel = "^0.0.22"



[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
